added fixing types sizes/byte_offsets to uint32_t and element_indexes and tuple_indexes to uint16_t

+++ inc/tuple_def.h
@@ -62,19 +62,19 @@ struct element_def
 
 	// size in bytes that may be occupied by the element
 	// size may be equal to VARIABLE_SIZED only for STRING and BLOB datatypes
-	uint64_t size;
+	uint32_t size;
 
 	// byte offset in tuple for the given element definition
 	// valid only if the tuple is not VARIABLE_SIZED
-	uint64_t byte_offset;
+	uint32_t byte_offset;
 };
 
 // check if a given size is valid for an element's datatype
-int is_size_allowed(element_type ele_type, uint64_t size);
+int is_size_allowed(element_type ele_type, uint32_t size);
 
 // initialize an element's definition using its type and size
 // it may fail if the size parameters is not valid for a given data type
-int init_element_def(element_def* element_d, element_type ele_type, uint64_t size);
+int init_element_def(element_def* element_d, element_type ele_type, uint32_t size);
 
 // compare 2 elements, given their element definition
 // this function must be called only on fixed sized elements i.e. ele_d->size != VARIABLE_SIZED
@@ -86,10 +86,10 @@ struct tuple_def
 {
 	// total size of tuple in bytes
 	// size = VARIABLE_SIZED, if atleast one of the element_defs is VARIABLE_SIZED
-	uint64_t size;
+	uint32_t size;
 
 	// total elements in the tuple
-	uint64_t element_count;
+	uint16_t element_count;
 
 	// definition of all of the elements
 	element_def element_defs[];
@@ -100,7 +100,7 @@ void init_tuple_def(tuple_def* tuple_d);
 
 // insert the key or values, insert keys in their decreasing order of importance
 // mark the tuple_mark_key_complete, once all the keys are inserted
-int insert_element_def(tuple_def* tuple_d, element_type ele_type, uint64_t element_size);
+int insert_element_def(tuple_def* tuple_d, element_type ele_type, uint32_t element_size);
 
 // after inserting all the elements call this function
 void finalize_tuple_def(tuple_def* tuple_d);
